#49 Made MultiplexedDatagramEndpoint more independent from ID
The ServerMultiplexedDatagramEndpoint is now completely independent from
the multiplexing ID (all checking and preparation is done in the
DatagramEndpointMultiplexer, the DatagramMessageWithIdSender and the
MultiplexingReceiver). This also allowed to remove the duplicate check
for a correct MultiplexingID.

The ClientMultiplexedDatagramEndpoint still has to check the ID on
arrival of a message (on the server this is done by the
MultiplexingReceiver) and therefore has to know the MultiplexingID.
Sending works the same way as on the server.