Imported CA certificate into a Java-readable keystore
The standard java classes have their own format for storing certificates, which
can be managed by the "keytool" command. The "keystore" file consists of the
CA certificate, which is marked as trusted, and is supposed to be used by the
client for verifying the certificate chain. The password is "bifrost" (keytool
does not allow for files without passwords).