First implementation of signature converters.
Datagram sending and receiving converters that authenticate messages
based on digital signatures. The server has a keystore containing a
certificate chain along with a private key. The client has a keystore
with the root certificate.

The current implementation is only a first attempt and needs
considerable refactoring, for instance, we should introduce an
intermediate abstraction abstracting from the actual scheme that is
used.